bug hunting

+++ src/org/poly2tri/Trapezoid.scala
@@ -74,7 +74,7 @@ class Trapezoid(val leftPoint: Point, var rightPoint: Point, val top: Segment, v
   }
   
   def lineIntersect(s: Segment, x: Float) = {
-    val y =  s.slope * x + s.b
+    val y =  (s.slope * x + s.b).toFloat
     Point(x, y)
   }
